Add SGX random engine
