+++ make/contrib/random.mk
@@ -1,4 +1,4 @@
-RANDOM_CONTRIB_SRC = $(wildcard src/contrib/random/*.cc)
+RANDOM_CONTRIB_SRC = $(wildcard src/contrib/random/random.cc)
 RANDOM_CONTRIB_OBJ = $(patsubst src/%.cc, build/%.o, $(RANDOM_CONTRIB_SRC))
 
 ifeq ($(USE_RANDOM), 1)
